requirements.txt
argcomplete==1.12.2
boto==2.49.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
crcmod==1.7
cryptography==3.4.7
fasteners==0.16
gcs-oauth2-boto-plugin==2.7
google-api-core==1.26.2
google-apitools==0.5.31
google-auth==1.28.0
google-cloud-core==1.6.0
google-cloud-storage==1.37.0
google-crc32c==1.1.2
google-reauth==0.1.1
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
gsutil==4.60
httplib2==0.19.0
idna==2.10
mock==2.0.0
monotonic==1.5
oauth2client==4.1.3
packaging==20.9
pbr==5.5.1
prettytable==2.1.0
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytz==2021.1
pyu2f==0.1.5
requests==2.25.1
retry-decorator==1.1.1
rsa==4.7.2
six==1.15.0
urllib3==1.26.4
wcwidth==0.2.5
